remove refactoring of pallets
