+++ runtime/common/src/lib.rs
@@ -17,4 +17,3 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 mod apis;
-mod pallets;
